updates to namespaces
