+++ src/exceptions/ClassMethodNotFound.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace obray;
+
+Class ClassMethodNotFound extends \Exception
+{
+
+}
