+++ src/exceptions/ClassNotFound.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace obray;
+
+Class ClassNotFound extends \Exception
+{
+
+}
