+++ src/interfaces/oUserInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * @license http://www.opensource.org/licenses/mit-license.php MIT (see the LICENSE file)
+ */
+
+namespace obray;
+
+/**
+ * Describes the interface of an obray users
+ */
+interface oUsersInterface
+{
+        public function checkPermissions(mixed $obj, bool $direct);
+}
